requirements.txt
cvu-python
vidsz
